+++ .gitattributes
@@ -1,0 +1,5 @@
+*.js    eol=lf
+*.jsx   eol=lf
+*.json  eol=lf
+*.ts    eol=lf
+*.tsx   eol=lf
